Module rstml::node::atoms

Tokens that is used as parts of nodes, to simplify parsing. Example: <!-- <> </> <! />

Also contain some entities that split parsing into several small units, like: <open_tag attr /> </close_tag>

Structs

CloseTag

Open tag for element, <name attr=x, attr_flag>

CloseTagStart

Start part of element’s close tag. Its commonly used as separator </

ComEnd

End part of comment tag -->

ComStart

Start part of comment tag <!--

Dash

DocStart

Start part of doctype tag <!

FragmentClose

Fragment close part </>

FragmentOpen

Fragment open part <>

OpenTag

Open tag for element, possibly self-closed. <name attr=x, attr_flag>

OpenTagEnd

End part of element’s open tag /> or >